Ensure coachmark layers behind header on example page (#7607)
On the example page for the coachmark we can see the coachmark appearing above the header if we scroll down far enough. This change makes the coachmark layer on the same layer as the button it is attached to, ensuring that it displays on the same layer underneath the floating header.